Refactor incremental lockout
Removed lockout_count (failed_logins redundancy)
Removed lockout_context function (just read locked_until)
Move failed_login logic into a model method
